Updates to Zipkin 2.16 by vendoring internal code previously borrowed
Before, this module used some internal types relating to okhttp client,
used in elasticsearch. This copies in the code as it is no-longer used
in Zipkin. This also copies in tests that pass from the last version in
Zipkin.

What happened was Zipkin changed to use the netty-based Armeria client,
so these okhttp types will no longer be maintained.